move mockClipboardAPI into its own function
